refactor: moved all resources into proper directories
