+++ src/app/app-routing.module.ts
@@ -1,17 +1,17 @@
-import {LoginComponent} from './pages/login/login.component';
-import {NgModule} from '@angular/core';
-import {RouterModule, Routes} from '@angular/router';
-import {UserListComponent} from "./user-list/user-list.component";
-import {PostListComponent} from "./post-list/post-list.component";
-import {AddPostComponent} from "./add-post/add-post.component";
-import {AuthGuard} from "./services/AuthGuard";
-import { ProfileComponent } from './pages/profile/profile.component';
-import {RegisterComponent} from "./pages/register/register.component";
+import { LoginComponent } from './login/login.component';
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { AddPostComponent } from './add-post/add-post.component';
+import { ProfileComponent } from './profile/profile.component';
+import { RegisterComponent } from './register/register.component';
+import { UserListComponent } from './user-list/user-list.component';
+import { PostListComponent } from './post-list/post-list.component';
+import { AuthGuard } from 'src/services/AuthGuard';
 
 const routes: Routes = [
   {
     path: '',
-    component: LoginComponent
+    component: LoginComponent,
   },
   {
     path: 'login',
@@ -24,31 +24,27 @@ const routes: Routes = [
   {
     path: 'users',
     component: UserListComponent,
-    canActivate: [AuthGuard]
+    canActivate: [AuthGuard],
   },
   {
     path: 'posts',
     component: PostListComponent,
-    canActivate: [AuthGuard]
+    canActivate: [AuthGuard],
   },
   {
     path: 'add-post',
     component: AddPostComponent,
-    canActivate: [AuthGuard]
+    canActivate: [AuthGuard],
   },
   {
     path: 'profile/:username',
     component: ProfileComponent,
-    canActivate: [AuthGuard]
+    canActivate: [AuthGuard],
   },
 ];
 
-
 @NgModule({
-  imports: [
-    RouterModule.forRoot(routes)
-  ],
+  imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule],
 })
-export class AppRoutingModule {
-}
+export class AppRoutingModule {}
