+++ src/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppComponent } from './app.component';
-import { LoginComponent } from './pages/login/login.component';
+import { LoginComponent } from './login/login.component';
 import { MaterialModule } from './material.module';
 import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
@@ -10,7 +10,7 @@ import { UserListComponent } from './user-list/user-list.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { PostListComponent } from './post-list/post-list.component';
 import { AddPostComponent } from './add-post/add-post.component';
-import { TimeAgoPipe } from './time-ago-pipe/time-ago-pipe.pipe';
+import { TimeAgoPipe } from '../pipes/time-ago-pipe.pipe';
 import { MatButtonModule } from '@angular/material/button';
 import { MatInputModule } from '@angular/material/input';
 import { MatCardModule } from '@angular/material/card';
@@ -20,9 +20,9 @@ import { MatListModule } from '@angular/material/list';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CommonModule } from '@angular/common';
 import { MatLineModule } from '@angular/material/core';
-import { ProfileComponent } from './pages/profile/profile.component';
+import { ProfileComponent } from './profile/profile.component';
 import { MatDividerModule } from '@angular/material/divider';
-import { RegisterComponent } from './pages/register/register.component';
+import { RegisterComponent } from './register/register.component';
 import { MatSelectModule } from '@angular/material/select';
 import { MatMenuModule } from '@angular/material/menu';
 import { ApiHttpInterceptor } from 'src/interceptors/http.interceptor';
