--- src/app/models/Post.ts
@@ -1,7 +1,0 @@
-export interface Post {
-  title: string;
-  content: string;
-  category: string;
-  likes: number;
-  publicationDate: string;
-}
