+++ src/app/register/register.component.ts
@@ -1,11 +1,11 @@
 import { Component } from '@angular/core';
-import {UserService} from "../../services/user.service";
-import {Router} from "@angular/router";
+import { Router } from '@angular/router';
+import { UserService } from 'src/services/user.service';
 
 @Component({
   selector: 'app-register',
   templateUrl: './register.component.html',
-  styleUrls: ['./register.component.scss']
+  styleUrls: ['./register.component.scss'],
 })
 export class RegisterComponent {
   user: any = {};
@@ -15,7 +15,7 @@ export class RegisterComponent {
   async register(): Promise<void> {
     try {
       await this.userService.registerUser(this.user);
-      await this.router.navigate(["/login"]);
+      await this.router.navigate(['/login']);
     } catch (error) {
       console.error('Registration failed!', error);
     }
