--- src/app/user.model.ts
@@ -1,8 +1,0 @@
-export interface User {
-  name: string;
-  email: string;
-  gender: string;
-  age: number;
-  username: string;
-  password: string;
-}
